Refactor VMMonitor_module.cc to move VMInfo to GeneralUtilties.
